+++ assets/main.scss
@@ -26,7 +26,7 @@ body {
     background-image: url("../../CorsaceAssets/img/ayim-mca/site/bg.png");
     background-position: center;
     background-repeat: no-repeat;
-    background-size: auto;
+    background-size: cover;
     font-family: 'Red Hat Display', sans-serif;
     color: white;
 	font-size: 1rem;
