+++ vue-shim.d.ts
@@ -1,0 +1,5 @@
+declare module "*.vue" {
+    import Vue from "vue";
+    import "@nuxt/types";
+    export default Vue;
+}
